introduce Reduction type tasks
While Serial tasks input N items and output M items, performing
some operation on them in a serial mannner, and Parallel tasks
input N *lists of* items and output M *lists of* items, mapping
an operation onto (sets of) their elements, we lack an operation
that would reduce a *list of* items into a single item.
This commit introduces Reduction tasks. They are allowed to only
input 1 list of items and output 1 item (extension will be looked
into). These tasks behave as both Parallel and Serial tasks, in
the following way:
* when a Reduction task is created (added to the flow), it is
  added as (kind of) a Parallel task - it is added to a MicroFlow,
  which then executes it in parallel, like any other of its tasks;
* upon completion of the MicroFlow sequence, the output from a
  Parallel task is forcibly gathered and piped back to MacroFlow;
* but a Reduction task is *also* added to the MacroFlow as another
  Serial task - this time to process the outputs of reduction in
  each subprocess.
For this reason, Reduction tasks must be invariant to the number
of elements in the original list. For example, precise calculation
of a mean is impossible in general: consider a case of a list of
numbers whose length is an odd integer. If this list is mapped
onto an even number of subprocesses, each of the arising sub-lists
will have a different number of elements. Averaging the sub-list
averages will *not* be exactly equal the average of the original.
The main limitation behind this is the mechanism of passing data
between the tasks, which does not include any metadata to allow
correcting for that. Also the intended simplicity of the task
definitions makes this difficult - precise mean calculation would
require the user to define two stages of the operation: first
calculating the sub-list sums and counts, then the weighted mean
of these sub-lists. Current, simple implementation only requires
defining a single operation.
Working around this problem is possible: mean can be computed as
first reducing the list with a "sum" operation, and then dividing
the output by the element count of the list.

There are some technical caveats in the Reduction base class. Four
functions are defined to accomodate execution as a Parallel as
well as a Serial task. The user defines further three functions to
describe their intended behavior from initialization to completion.
The reduction mechanism is defined as an operation r(a,b), which
describes how to reduce any two items. The list of N items is then
reduced to 1 item by reducing the first 2 items into an accumulator
variable, and then iterating over the remaining items, reducing
every one into the accumulator again. The value of the accumulator
after the loop is the result of reduction. These steps are realized
within the framework in a following way:
* first, the Reducer task is initialized with the setup() function;
* then it is copied into the subprocesses;
* at first, the "action" function is set to the "action_first",
  which initializes the accumulator with the first value;
* then the "action" is rebound to the "action_every", which will be
  called at every iteration over the input list;
* "action_every" calls "reduction" on the accumulator and the next
  data element; "reduction" is the user-defined operation r();
* eventually, "output" is called, which calls the user-defined
  "final" to process the last value of the accumulator and returns;
* after the Parallel part, the list of N elements has been reduced
  to M elements, where M is the number of processes; this is then
  reduced to a single element using a Serial version;
* when MacroFlow instantiates a Serial version of Reduction task,
  it rebinds its "action" function to "action_final";
* when this is called, it performs all the above steps: initializes
  the buffer, iterates over the elements, calls "final" and returns
  the resulting value.