script parsing now done by a separate Parser object
Significant cleanup of the parsing mechanism, with the whole burden
moved to a dedicated object. BaseProcessor no longer bothers with
text processing - it receives a dict containing lists of inputs,
outputs and parameters. It's only job is now to verify these lists
with respect to its class specification. Role of the Assembler is
also minimized - now it only calls Parser to process the text into
a dict that the task ingests directly.