adopt a metaclass for all tasks to derive from
Metaclass will set the isValid flag to False for every class that
derives from it. This fixes the problem of deriving classes being
excluded from parameter validation by Assembler (since after base
classes were validated, and their isValid flags set to true, the
derived ones also inherit the truthy flag).
Also, it allows marking any base classes with a special flag that
is not inherited (or rather, it is set to False to any inheriting
class, unless explicitly stated in the class definition). Because
of this, Assembler does not have to check what instance the given
class is (before deciding whether to process it or not) - instead
simply checking the attribute. As an additional benefit, user can
define their own base class and mark it as such, thus preventing
it from being added to the Assembler.
Because metaclasses have different syntax in Python 2 and 3, this
commit introduces a compatibility layer to allow the mechanism in
either version (see metaClass2.py, metaClass3.py & metaBase.py).