allow MicroFlow to directly query MacroFlow for its scope
This way it's easier for Macro to query all existing Micros for
their outputs and gather them on the fly. The most important
benefit of this change is that an item produced by one MicroFlow
but NOT used by any serial task can now be directly reached by
another MicroFlow.